+++ src/transforms/serializer.js
@@ -4,6 +4,17 @@ var Parser = require("protodef").Parser;
 
 var minecraft = require("../datatypes/minecraft");
 
+
+function recursiveAddTypes(protocol, protocolData, path) {
+  if(protocolData === undefined)
+    return
+
+  if(protocolData.types)
+    protocol.addTypes(protocolData.types);
+
+  recursiveAddTypes(protocol,get(protocolData,path.shift()),path);
+}
+
 function createProtocol(packets) {
   var proto = new ProtoDef();
 
@@ -17,6 +28,7 @@ function createSerializer(isServer = false) {
   var direction = !isServer ? 'toServer' : 'toClient';
   var packets = mcData[direction].types;
   var proto = createProtocol(packets);
+  recursiveAddTypes(proto, merge(mcData.protocol,get(customPackets,[mcData.version.majorVersion])),[state,direction]);
   return new Serializer(proto, "packet");
 }
 
@@ -25,6 +37,7 @@ function createDeserializer(isServer = false) {
   var direction = isServer ? "toServer" : "toClient";
   var packets = mcData[direction].types;
   var proto = createProtocol(packets);
+  recursiveAddTypes(proto,merge(mcData.protocol,get(customPackets,[mcData.version.majorVersion])),[state,direction]);
   return new Parser(proto, "packet");
 }